requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
elevenlabs==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
PyAudio==0.2.14
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1

# Production dependencies:
fastapi
uvicorn
python-multipart

# Development dependencies:
pytest
